refactor: separating JacProgram out
