+++ jac/scripts/check.sh
@@ -1,0 +1,1 @@
+pre-commit run --files $(find . -type f)
